src/requirements.txt
discord.py
aiosqlite
python-dotenv
requests
